fixup! code review
