+++ apps/site/components/Downloads/DownloadLink.tsx
@@ -5,7 +5,7 @@ import type { FC, PropsWithChildren } from 'react';
 import LinkWithArrow from '#site/components/LinkWithArrow';
 import { useClientContext } from '#site/hooks';
 import type { DownloadKind, NodeRelease } from '#site/types';
-import { getNodeDownloadUrl } from '#site/util/download';
+import { getNodeDownloadUrl } from '#site/util/url';
 import { getUserPlatform } from '#site/util/userAgent';
 
 type DownloadLinkProps = { release: NodeRelease; kind?: DownloadKind };
